Editted impls

+++ java-school-fullstack/src/main/java/mthree/com/fullstackschool/service/StudentServiceImpl.java
@@ -32,7 +32,17 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public Student getStudentById(int id) {
         //YOUR CODE STARTS HERE
 
-        return this.studentDao.findStudentById(id);
+        Student student = new Student();
+        try
+        {
+            student = this.studentDao.findStudentById(id);
+        }
+        catch (DataAccessException e)
+        {
+            student.setStudentFirstName("Student Not Found");
+            student.setStudentLastName("Student Not Found");
+        }
+        return student;
 
         //YOUR CODE ENDS HERE
     }
@@ -40,7 +50,19 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public Student addNewStudent(Student student) {
         //YOUR CODE STARTS HERE
 
-        return this.studentDao.createNewStudent(student);
+        boolean fieldIsEmpty = false;
+        if (student.getStudentFirstName().isEmpty())
+        {
+            fieldIsEmpty = true;
+            student.setStudentFirstName("First Name blank, student NOT added");
+        }
+
+        if (student.getStudentLastName().isEmpty())
+        {
+            fieldIsEmpty = true;
+            student.setStudentLastName("Last Name blank, student NOT added");
+        }
+        return fieldIsEmpty ? student : this.studentDao.createNewStudent(student);
 
         //YOUR CODE ENDS HERE
     }
@@ -48,10 +70,19 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public Student updateStudentData(int id, Student student) {
         //YOUR CODE STARTS HERE
 
+        /*
         Student studentToUpdate = this.studentDao.findStudentById(id);
         if (studentToUpdate == null)
         {
             return null;
+        }
+         */
+
+        if (id != student.getStudentId())
+        {
+            student.setStudentFirstName("IDs do not match, student not updated");
+            student.setStudentLastName("IDs do not match, student not updated");
+            return student;
         }
         this.studentDao.updateStudent(student);
         return this.studentDao.findStudentById(id);
@@ -70,7 +101,16 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public void deleteStudentFromCourse(int studentId, int courseId) {
         //YOUR CODE STARTS HERE
 
-        this.studentDao.addStudentToCourse(studentId, courseId);
+        Student student = studentDao.findStudentById(studentId);
+        if (student.getStudentFirstName().equals("Student Not Found"))
+        {
+            System.out.println("Student Not Found");
+        }
+        else
+        {
+            this.studentDao.addStudentToCourse(studentId, courseId);
+            System.out.println("Student:" + studentId + " deleted from course:" + courseId);
+        }
 
         //YOUR CODE ENDS HERE
     }
@@ -78,7 +118,23 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public void addStudentToCourse(int studentId, int courseId) {
         //YOUR CODE STARTS HERE
 
-        this.studentDao.deleteStudentFromCourse(studentId, courseId);
+        Student student = studentDao.findStudentById(studentId);
+        if (student.getStudentFirstName().equals("Student Not Found"))
+        {
+            System.out.println("Student Not Found");
+        }
+        else
+        {
+            try
+            {
+                this.studentDao.deleteStudentFromCourse(studentId, courseId);
+                System.out.println("Student:" + studentId + " added to course:" + courseId);
+            }
+            catch(Exception e)
+            {
+                System.out.println("Student:" + studentId + " already enrolled on course:" + courseId);
+            }
+        }
 
         //YOUR CODE ENDS HERE
     }
